pyproject.toml
[project]
name = "ai-a2a"
version = "0.1.0"
description = "A2A Demo"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13.3",
    "fastapi>=0.111.0",
    "httpx>=0.27.0",
    "litellm>=1.82.0",
    "minio>=7.2.20",
    "openai>=1.30.0",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.9",
    "qdrant-client>=1.9.0",
    "requests>=2.32.5",
    "solace-pubsubplus>=1.11.0",
    "uvicorn[standard]>=0.29.0",
]
